move commtrack settings view from the project settings to setup tab
